Minor cosmetic changes
