+++ src/TestApps.Tests/WpfTestApplication.Tests/WpfTestApplication.Tests.csproj
@@ -47,14 +47,15 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AdditionalCommandTests\TestWebDriver.cs" />
+    <Compile Include="TestHelper.cs" />
+    <Compile Include="TestWebDriver.cs" />
     <Compile Include="BaseForMainWindowTest.cs" />
     <Compile Include="BaseTest.cs" />
     <Compile Include="ActionChainsTests.cs" />
     <Compile Include="CommandTests\ClearElementTests.cs" />
     <Compile Include="CommandTests\ClickElementTests.cs" />
     <Compile Include="CommandTests\GetActiveElementTests.cs" />
-    <Compile Include="AdditionalCommandTests\DataGridCommandsTests.cs" />
+    <Compile Include="WiniumCommandTests\DataGridCommandsTests.cs" />
     <Compile Include="CommandTests\MouseClickTests.cs" />
     <Compile Include="CommandTests\MouseDoubleClickTests.cs" />
     <Compile Include="CommandTests\QuitTests.cs" />
@@ -77,12 +78,6 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\WindowsFormsTestApplication.Tests\WindowsFormsTestApplication.Tests.csproj">
-      <Project>{DC69009A-39E7-4F02-B721-3CC99B4EA637}</Project>
-      <Name>WindowsFormsTestApplication.Tests</Name>
-    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
